+++ Logic/Nodes/FeaturesNodes.cpp
@@ -144,7 +144,7 @@ public:
 		// Do stuff
 		std::vector<cv::Vec2f> linesVector;
 		cv::HoughLines(src, linesVector, _rhoResolution, CV_PI/180 * _thetaResolution, _threshold);
-		lines.create(linesVector.size(), 2, CV_32F);
+		lines.create(static_cast<int>(linesVector.size()), 2, CV_32F);
 
 		float* linesPtr = lines.ptr<float>();
 		for(const auto& line : linesVector)
